% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dex.R
\name{dex_sentence}
\alias{dex_sentence}
\title{Generate Random sentences of Pokémon using moves}
\usage{
dex_sentence(
  n,
  gen = 1:9,
  case = c("sentence", "snake", "small_camel", "big_camel", "screaming_snake", "parsed",
    "mixed", "lower_upper", "upper_lower", "swap", "all_caps", "lower_camel",
    "upper_camel", "internal_parsing", "none", "flip", "random", "title"),
  ...
)
}
\arguments{
\item{n}{An integer. Number of IDs to generate.}

\item{gen}{vector of integers. The generations of Pokemon to use. Defaults to all generations.}

\item{case}{Character string. The case style to use. Supports all styles implemented in the package \code{snakecase}.}

\item{...}{other arguments passed to \code{snakecase::to_any_case()}.}
}
\value{
A character vector of length \code{n} containing the generated sentences.
}
\description{
Generate Random sentences of Pokémon using moves
}
\examples{
dex_sentence(3)
dex_sentence(5)

}
