% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dex.R
\name{dexid}
\alias{dexid}
\title{Generate Random IDs Using Adjectives and Pokemon Names}
\usage{
dexid(
  n,
  n_adj = 1,
  gen = 1:9,
  case = c("snake", "kebab", "camel", "caps", "sentence", "custom"),
  add_digits = FALSE,
  ...
)
}
\arguments{
\item{n}{An integer. Number of IDs to generate.}

\item{n_adj}{An integer. Number of adjectives to use. Defaults to 1.}

\item{gen}{vector of integers. The generations of Pokemon to use. Defaults to all generations.}

\item{case}{Character string. The case style to use. Supported cases are \code{snake}, \code{kebab}, \code{camel}, \code{caps}, sentence and custom. If \code{case} is set to \code{custom}, a \code{delimiter} argument must be provided.}

\item{add_digits}{Logical. If \code{TRUE}, adds a random digit to the end of each ID. Defaults to \code{FALSE}.}

\item{...}{other arguments passed to \code{snakecase::to_any_case()}.}
}
\value{
A character vector of length \code{n} containing the generated IDs.
}
\description{
Creates unique identifiers by combining random adjectives with Pokemon names
in different formats.
}
\examples{
# Generate 5 basic snake_case IDs
dexid(5)

# Generate 5 IDs with 2 adjectives for more uniqueness
dexid(5, n_adj = 2)

# Generate 5 Title Case IDs
dexid(5, case = "title")

# Generate 5 IDs with random digits
dexid(5, add_digits = TRUE)

# Generate 5 IDs restrict to specific generation(s)
dexid(5, gen = 1)
}
